initial changes

+++ qf_lib/backtesting/execution_handler/commission_models/bps_trade_value_commission_model.py
@@ -29,7 +29,7 @@ class BpsTradeValueCommissionModel(CommissionModel):
     def __init__(self, commission: float):
         self.commission = commission
 
-    def calculate_commission(self, fill_quantity: int, fill_price: float) -> float:
+    def calculate_commission(self, fill_quantity: float, fill_price: float) -> float:
         fill_quantity = abs(fill_quantity)
         commission = fill_price * fill_quantity * self.commission / 10000
         return commission
